+++ app/application/create/index.tsx
@@ -9,17 +9,16 @@ import BackArrow from "@/src/components/ui/BackArrow";
 import Button from "@/src/components/ui/Button/Button";
 import ImagePicker from "@/src/components/ui/ImagePicker";
 import TextSize from "@/src/components/ui/TextSize";
-import { useApplicationFormValidation } from "@/src/hooks/application/useApplicationFormValidation";
+import { useApplicationFormValidation } from "@/src/hooks/application/create/useApplicationFormValidation";
 import { useApplicationSlice } from "@/src/stores/slices/useApplicationSlice";
 import { formatPhoneNumber } from "@/src/utils/common/phoneNumberFormatter";
-import { COLORS } from "@/src/utils/constants/colors";
 import { GENDER_BASIC } from "@/src/utils/constants/options";
+import TextInput from "../../../src/components/ui/TextInput";
 import DateInput from "../_components/DateInput";
 import FormField from "../_components/FormField";
 import FormSection from "../_components/FormSection";
 import GenderSelector from "../_components/GenderSelector";
 import ProgressBar from "../_components/ProgressBar";
-import TextInput from "../_components/TextInput";
 
 export default function applicationCreate() {
   const {
@@ -158,7 +157,7 @@ export default function applicationCreate() {
                 </FormSection>
 
                 <Button
-                 variant='primary'
+                  variant="primary"
                   width={370}
                   height={50}
                   textColor="white"
