+++ src/hooks/application/create/useProgressAnimation.ts
